添加了程序运行时非处理异常的处理逻辑，修复了未找到配置文件抛出异常的问题。

+++ TaskSchedule.cs
@@ -124,7 +124,10 @@ namespace AutoDarkModeMin
             }
             catch (Exception e)
             {
-                File.Delete("setting.json");
+                if (File.Exists(appFolder + "\\setting.json"))
+                {
+                    File.Delete("setting.json");
+                }
             }
 
         }
